bugfix: commented out @builder annotation in some users, relocated resourcecomponent files to respective folders entity, repository, service, controller
