+++ src/main/java/pl/io/emergency/entity/ResourceType.java
@@ -1,4 +1,4 @@
-package pl.io.emergency.resource;
+package pl.io.emergency.entity;
 
 public enum ResourceType {
     CLOTHES,
