+++ src/main/java/pl/io/emergency/entity/Volunteer.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
+//@Builder
 @EqualsAndHashCode(callSuper = true)
 @Entity
 public class Volunteer extends User {
