+++ src/main/java/pl/io/emergency/repository/ResourceRepositorium.java
@@ -1,9 +1,9 @@
-package pl.io.emergency.resource;
+package pl.io.emergency.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.io.emergency.resource.ResourceEntity;
+import pl.io.emergency.entity.ResourceEntity;
 
 @Repository
 public interface ResourceRepositorium extends JpaRepository<ResourceEntity, Long> {
