+++ src/test/java/pl/io/emergency/resource/ResourceDtoControllerTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.ResponseEntity;
-import pl.io.emergency.resource.ResourceController;
-import pl.io.emergency.resource.ResourceDto;
-import pl.io.emergency.resource.ResourceType;
-import pl.io.emergency.resource.ResourceStatus;
+import pl.io.emergency.controller.ResourceController;
+import pl.io.emergency.entity.ResourceDto;
+import pl.io.emergency.entity.ResourceStatus;
+import pl.io.emergency.entity.ResourceType;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
