+++ src/test/java/pl/io/emergency/resource/ResourceEntityTest.java
@@ -2,8 +2,9 @@ package pl.io.emergency.resource;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-
-import java.util.Optional;
+import pl.io.emergency.entity.ResourceEntity;
+import pl.io.emergency.entity.ResourceType;
+import pl.io.emergency.repository.ResourceRepositorium;
 
 import static org.junit.jupiter.api.Assertions.*;
 public class ResourceEntityTest {
