+++ src/test/java/pl/io/emergency/resource/ResourceServiceTest.java
@@ -2,13 +2,17 @@ package pl.io.emergency.resource;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
+import pl.io.emergency.entity.ResourceDto;
+import pl.io.emergency.entity.ResourceEntity;
+import pl.io.emergency.entity.ResourceStatus;
+import pl.io.emergency.entity.ResourceType;
+import pl.io.emergency.repository.ResourceRepositorium;
+import pl.io.emergency.service.ResourceService;
 
 import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.when;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
 
 public class ResourceServiceTest {
     @Test
